requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.9
django-celery-results==2.2.0
dnspython==2.1.0
eventlet==0.33.0
greenlet==1.1.2
idna==3.3
kombu==5.2.2
prompt-toolkit==3.0.22
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5

djangorestframework~=3.12.4